app/requirements.txt
grpcio-tools==1.64.1
grpcio==1.64.1
psycopg2==2.9.9